+++ src/core/lifecycle.ts
@@ -21,6 +21,21 @@ function dispose(hooks: TeardownHook[]) {
 	}
 }
 
+export type LeakHook = (hook: TeardownHook) => void;
+
+/**
+ * Register a hook to be called when any teardown hooks are registered outside of any capture calls.
+ *
+ * Errors thrown from the leak hook will be thrown by the **teardown** calls.
+ */
+export function onLeak(hook: LeakHook): void {
+	if (TEARDOWN_STACK.length > 0) {
+		// onLeak must only be called once and outside of any capture calls:
+		throw new Error("G4");
+	}
+	TEARDOWN_STACK.push({ push: hook });
+}
+
 /**
  * Run a function while capturing teardown hooks.
  *