+++ src/test/index.ts
@@ -1,4 +1,3 @@
-export * from "./lifecycle.js";
 export * from "./poll.js";
 export * from "./queue.js";
 export * from "./run-async.js";