+++ tests/env-rvxdom.tsx
@@ -1,9 +1,8 @@
-import { ENV } from "rvx";
+import { ENV, onLeak } from "rvx";
 import { WINDOW } from "rvx/dom";
-import { onTeardownLeak } from "rvx/test";
 
 ENV.default = WINDOW;
 
-onTeardownLeak(() => {
+onLeak(() => {
 	throw new Error("teardown leak");
 });